chore(property): decouple set/get flows
This changes the behavior such that we're not doing a full roundtrip of
getting and returning the property after setting it.

Passing the input value through to the output in the "set" flow seems
unnecessary and would require cloning the value, so I opted not to do
that. We have to put _something_ in the state/output though, so I'm
using a JSON null as a placeholder.